Deleted task folder. Simplified test. Removed unused code.
