+++ packages/primitive-contracts/package.json
@@ -22,11 +22,13 @@
         "venv": "virtualenv .venv",
         "slither": "pip install slither-analyzer && slither .",
         "bevm": "buidler node",
-        "test": "cross-env buidler test ./test/option/trader.js ./test/option/option.js",
-        "test:local": "cross-env buidler test --network local ./test/option/trader.js ./test/option/option.js",
+        "test": "cross-env buidler test ./test/trader.js ./test/option.js ./test/uni-trader.js",
+        "test:local": "cross-env buidler test --network local ./test/trader.js ./test/option.js ./test/uni-trader.js",
         "deploy:local": "cross-env buidler deploy --network local",
         "deploy:rinkeby": "cross-env buidler deploy --network rinkeby",
         "deploy:mainnet": "cross-env buidler deploy --network live",
+        "deploy-options:rinkeby": "cross-env run ./scripts/deploy_options.js --network rinkeby",
+        "deploy-options:mainnet": "cross-env run ./scripts/deploy_options.js --network live",
         "verify:rinkeby": "cross-env buidler run ./scripts/verify.js --network rinkeby",
         "verify:mainnet": "cross-env buidler run ./scripts/verify.js --network live"
     },
