+++ packages/primitive-contracts/scripts/deploy_options.js
@@ -30,44 +30,34 @@ async function checkAllowance(owner, spender, token) {
     }
 }
 
+/* const deployOption = async () => {
+    const { registry, optionFactory, redeemFactory } = await setupRegistry();
+    const { usdcToken, ethToken } = await setupTokens();
+    const base = parseEther("1");
+    const quote = parseEther("1000");
+    const expiry = "1609286400";
+    await checkSupported(registry, ethToken, usdcToken);
+    await checkInitialization(registry, optionFactory, redeemFactory);
+    const tx = await registry.deployOption(ethToken.address, usdcToken.address, base, quote, expiry, { gasLimit: 1000000 });
+    const deployedOption = await registry.getOption(ethToken.address, usdcToken.address, base, quote, expiry);
+
+    return { tx, deployedOption };
+}; */
+
 const deployOption = async () => {
     const [signer] = await ethers.getSigners();
     const account = await signer.getAddress();
-    const registry = new ethers.Contract(
-        Registry.address,
-        Registry.abi,
-        signer
-    );
-    const optionFactory = new ethers.Contract(
-        OptionFactory.address,
-        OptionFactory.abi,
-        signer
-    );
-    const redeemFactory = new ethers.Contract(
-        RedeemFactory.address,
-        RedeemFactory.abi,
-        signer
-    );
+    const registry = new ethers.Contract(Registry.address, Registry.abi, signer);
+    const optionFactory = new ethers.Contract(OptionFactory.address, OptionFactory.abi, signer);
+    const redeemFactory = new ethers.Contract(RedeemFactory.address, RedeemFactory.abi, signer);
     const trader = new ethers.Contract(Trader.address, Trader.abi, signer);
-    const uniswapTrader = new ethers.Contract(
-        UniswapTrader.address,
-        UniswapTrader.abi,
-        signer
-    );
+    const uniswapTrader = new ethers.Contract(UniswapTrader.address, UniswapTrader.abi, signer);
 
     const usdcToken = new ethers.Contract(USDC.address, USDC.abi, signer);
     const ethToken = new ethers.Contract(ETH.address, ETH.abi, signer);
 
-    const uniswapFactory = new ethers.Contract(
-        RINKEBY_UNI_FACTORY,
-        UniswapV2Factory.abi,
-        signer
-    );
-    const uniswapRouter = new ethers.Contract(
-        RINKEBY_UNI_ROUTER02,
-        UniswapV2Router02.abi,
-        signer
-    );
+    const uniswapFactory = new ethers.Contract(RINKEBY_UNI_FACTORY, UniswapV2Factory.abi, signer);
+    const uniswapRouter = new ethers.Contract(RINKEBY_UNI_ROUTER02, UniswapV2Router02.abi, signer);
 
     // approve the router
     await checkAllowance(account, uniswapRouter, usdcToken);
@@ -103,43 +93,21 @@ const deployOption = async () => {
         // check initialized and supported
         await checkSupported(registry, ethToken, usdcToken);
         await checkInitialization(registry, optionFactory, redeemFactory);
-        let deployedOption = await registry.getOption(
-            underlying,
-            quoteToken,
-            base,
-            quote,
-            expiry
-        );
+        let deployedOption = await registry.getOption(underlying, quoteToken, base, quote, expiry);
         // deploy an option
         let tx;
         if (deployedOption == ZERO_ADDRESS) {
             try {
-                tx = await registry.deployOption(
-                    underlying,
-                    quoteToken,
-                    base,
-                    quote,
-                    expiry,
-                    { gasLimit: 1000000 }
-                );
+                tx = await registry.deployOption(underlying, quoteToken, base, quote, expiry, { gasLimit: 1000000 });
             } catch (err) {
                 console.log(err);
             }
             // get deployed option address
-            deployedOption = await registry.getOption(
-                underlying,
-                quoteToken,
-                base,
-                quote,
-                expiry
-            );
+            deployedOption = await registry.getOption(underlying, quoteToken, base, quote, expiry);
         }
 
         // create a new pair
-        let pairAddress = await uniswapFactory.getPair(
-            deployedOption,
-            usdcToken.address
-        );
+        let pairAddress = await uniswapFactory.getPair(deployedOption, usdcToken.address);
 
         if (pairAddress == ZERO_ADDRESS) {
             try {
@@ -148,35 +116,20 @@ const deployOption = async () => {
                 console.log(err);
             }
 
-            pairAddress = await uniswapFactory.getPair(
-                deployedOption,
-                usdcToken.address
-            );
+            pairAddress = await uniswapFactory.getPair(deployedOption, usdcToken.address);
         }
 
-        let uniswapPair = new ethers.Contract(
-            pairAddress,
-            UniswapV2Pair.abi,
-            signer
-        );
+        let uniswapPair = new ethers.Contract(pairAddress, UniswapV2Pair.abi, signer);
         let liquidity = await uniswapPair.getReserves();
         let reserve0 = liquidity._reserve0;
         if (reserve0 == 0) {
             // approve the router to take the option liquidity
-            let optionTokenInstance = new ethers.Contract(
-                deployedOption,
-                ERC20.abi,
-                signer
-            );
+            let optionTokenInstance = new ethers.Contract(deployedOption, ERC20.abi, signer);
             await checkAllowance(account, uniswapRouter, optionTokenInstance);
 
             // mint new options
             try {
-                await trader.safeMint(
-                    deployedOption,
-                    parseEther("100"),
-                    await signer.getAddress()
-                );
+                await trader.safeMint(deployedOption, parseEther("100"), await signer.getAddress());
             } catch (err) {
                 console.log(err);
             }
