--- packages/primitive-contracts/tasks/index.js
@@ -1,8 +1,0 @@
-require("@nomiclabs/buidler/config");
-const option = require("./option");
-const trader = require("./trader");
-
-Object.assign(module.exports, {
-    option,
-    trader,
-});
