--- packages/primitive-contracts/tasks/test-task.js
@@ -1,9 +1,0 @@
-const { task } = require("@nomiclabs/buidler/config");
-
-require("@nomiclabs/buidler/config");
-
-task("fake", "Tests a task").setAction(() => {
-    console.log("works");
-});
-
-module.exports = {};
