+++ packages/primitive-contracts/test/lib/utils.js
@@ -1,4 +1,22 @@
 const { assert } = require("chai");
+const bre = require("@nomiclabs/buidler/config");
+const { ethers } = require("ethers");
+const { parseEther } = ethers.utils;
+const { AddressZero } = ethers.constants;
+
+async function checkAllowance(owner, spender, token) {
+    const amount = parseEther("10000000000");
+    let allowance = await token.allowance(owner.address, spender.address);
+    if (allowance <= amount) {
+        await token.approve(spender.address, amount, { from: owner.address });
+    }
+}
+
+async function checkInitialization(registry, optionFactory, redeemFactory) {
+    const fac = await registry.optionFactory();
+    if (fac == AddressZero || fac != optionFactory.address)
+        await registry.initialize(optionFactory.address, redeemFactory.address);
+}
 
 const assertBNEqual = (actualBN, expectedBN, message) => {
     assert.equal(actualBN.toString(), expectedBN.toString(), message);
@@ -45,4 +63,6 @@ module.exports = {
     verifyOptionInvariants,
     getTokenBalances,
     getTokenBalance,
+    checkAllowance,
+    checkInitialization,
 };
